+++ modules/documentation/docs/changelog.md
@@ -2,10 +2,11 @@
 
 ## Next Release
 
+
+- \[ops\] Add persistent storage for Grafana
 - \[contracts\] Add `transactionHash` to withdrawal commitment
 - \[store\] Add `transferDefinition` to filter options on `getTransfers`
 - \[store\] Add `getWithdrawalCommitmentByTransactionHash` method
-
 
 ## 0.1.15-beta.12